refactor: 파일 위치 이동에 따른 변경
